Finalizing handling of PGN.
Removed use of 'person' from BookWord.xsl output for all grammatical
categories where it does not apply in order to separate use of '*' to
mean "any" and missing/null values to represent "not applicable".

Replaced use of integers with English words for 'person' in BookNorm.xsl
to make it self-explanatory and compatible with general English grammar
standards.